Componente header criado e organização de componentes
