+++ src/index.css
@@ -1,0 +1,4 @@
+body {
+    margin: 0;
+    font-family: Arial, sans-serif;
+}
